[orderex] Only share journal directory with container directory
Instead of sharing the whole customer directory with the container, this commit
only shares the journal directory. This means that the code repositories are in
the container directory. This was done to avoid issues with code repositories on
the host, including the .venv directory, and ruff being in the container
directory - sorry, didn't write down the exact issue.